// src/service/file/schemas/smartFolderResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { SmartFolderResultSmartFolderApiType } from './smartFolderResultSmartFolderApiType';
import type { SmartFolderResultRootType } from './smartFolderResultRootType';
import type { SmartFolderResultPublicState } from './smartFolderResultPublicState';

export interface SmartFolderResult {
  /** 스마트폴더 자료의 id */
  id: number;
  /** api 요청하는 구분(Photo:사진스마트폴더, EducationalData:자료스마트폴더, UserFolder:내폴더) */
  smartFolderApiType: SmartFolderResultSmartFolderApiType;
  /** 스마트폴더의 루트 종류(주로 작업) */
  rootType: SmartFolderResultRootType;
  /** 상위 스마트폴더 자료의 id */
  parentSmartFolderItemId: number;
  /** 폴더의 물리적 키. (DriveItem.key */
  driveItemKey: string;
  /**
   * 폴더의 폴더 뎁스.
   * @minimum 1
   * @maximum 10
   */
  depth: number;
  /**
   * 폴더 이름
   * @maxLength 255
   */
  name: string;
  /** 폴더의 공개 상태. */
  publicState: SmartFolderResultPublicState;
  /** 해당 스마트폴더에서 숨김여부 */
  isHidden: boolean;
  /** 소유자 계정 Id */
  ownerAccountId: number;
  /** 소유자 프로필 Id */
  ownerProfileId: number;
  /** 자료의 실제 생성/업로드 일 */
  driveItemCreatedAt: string;
  /** 사용자가 이름을 바꾸거나 변경할 수 있는지 여부. */
  userEditable: boolean;
  /**
   * 대표사진의 url (그리드 리스트용)
   * @nullable
   */
  thumbUrl?: string | null;
  /**
   * 내부 사용용 id
   * @nullable
   */
  internalId?: number | null;
  /**
   * 위젯을 숨길 수 있는지 여부
   * @nullable
   */
  hidable?: boolean | null;
}